backend/requirements.txt
# FastAPI核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# 缓存和任务队列
redis==5.0.1
celery==5.3.4
flower==2.0.1

# 医学影像处理
pydicom==2.4.3
SimpleITK==2.3.1
nibabel==5.1.0
Pillow==10.1.0

# AI/ML
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
numpy==1.26.2
opencv-python==4.8.1.78
onnxruntime==1.16.3

# 对象存储
minio==7.2.0
boto3==1.29.7

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具库
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
openpyxl==3.1.2

# 监控和日志
loguru==0.7.2
sentry-sdk==1.38.0
prometheus-client==0.19.0

